requirements.txt
# Core dependencies for AnomFL
numpy>=1.22.0
pandas>=1.2.0
torch>=1.9.0

# Data analysis and visualization
matplotlib>=3.0.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Jupyter and development
jupyter>=1.0.0
notebook>=6.0.0
ipython>=7.0.0
ipykernel>=6.0.0

# Additional utilities
pyyaml>=5.0.0
colorlog>=6.0.0